+++ phenolic_api/src/main/java/pt/ipb/phenolic/repos/MSFragmentRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import pt.ipb.phenolic.models.MSFragment;
 
 @Repository
-public interface MSFragmentRepository extends JpaRepository<MSFragment, Integer> {
+public interface MSFragmentRepository extends JpaRepository<MSFragment, Long> {
 }
